% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupssettings_objects.R
\name{Groups}
\alias{Groups}
\title{Groups Settings API Objects 
Lets you manage permission levels and related settings of a group.}
\usage{
Groups(allowExternalMembers = NULL, allowGoogleCommunication = NULL,
  allowWebPosting = NULL, archiveOnly = NULL, customFooterText = NULL,
  customReplyTo = NULL, defaultMessageDenyNotificationText = NULL,
  description = NULL, email = NULL, includeCustomFooter = NULL,
  includeInGlobalAddressList = NULL, isArchived = NULL,
  maxMessageBytes = NULL, membersCanPostAsTheGroup = NULL,
  messageDisplayFont = NULL, messageModerationLevel = NULL, name = NULL,
  primaryLanguage = NULL, replyTo = NULL,
  sendMessageDenyNotification = NULL, showInGroupDirectory = NULL,
  spamModerationLevel = NULL, whoCanAdd = NULL, whoCanContactOwner = NULL,
  whoCanInvite = NULL, whoCanJoin = NULL, whoCanLeaveGroup = NULL,
  whoCanPostMessage = NULL, whoCanViewGroup = NULL,
  whoCanViewMembership = NULL)
}
\arguments{
\item{allowExternalMembers}{Are external members allowed to join the group}

\item{allowGoogleCommunication}{Is google allowed to contact admins}

\item{allowWebPosting}{If posting from web is allowed}

\item{archiveOnly}{If the group is archive only}

\item{customFooterText}{Custom footer text}

\item{customReplyTo}{Default email to which reply to any message should go}

\item{defaultMessageDenyNotificationText}{Default message deny notification message}

\item{description}{Description of the group}

\item{email}{Email id of the group}

\item{includeCustomFooter}{Whether to include custom footer}

\item{includeInGlobalAddressList}{If this groups should be included in global address list or not}

\item{isArchived}{If the contents of the group are archived}

\item{maxMessageBytes}{Maximum message size allowed}

\item{membersCanPostAsTheGroup}{Can members post using the group email address}

\item{messageDisplayFont}{Default message display font}

\item{messageModerationLevel}{Moderation level for messages}

\item{name}{Name of the Group}

\item{primaryLanguage}{Primary language for the group}

\item{replyTo}{Whome should the default reply to a message go to}

\item{sendMessageDenyNotification}{Should the member be notified if his message is denied by owner}

\item{showInGroupDirectory}{Is the group listed in groups directory}

\item{spamModerationLevel}{Moderation level for messages detected as spam}

\item{whoCanAdd}{Permissions to add members}

\item{whoCanContactOwner}{Permission to contact owner of the group via web UI}

\item{whoCanInvite}{Permissions to invite members}

\item{whoCanJoin}{Permissions to join the group}

\item{whoCanLeaveGroup}{Permission to leave the group}

\item{whoCanPostMessage}{Permissions to post messages to the group}

\item{whoCanViewGroup}{Permissions to view group}

\item{whoCanViewMembership}{Permissions to view membership}
}
\value{
Groups object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:54:49
filename: /Users/mark/dev/R/autoGoogleAPI/googlegroupssettingsv1.auto/R/groupssettings_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Groups Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
JSON template for Group resource
}
\seealso{
Other Groups functions: \code{\link{groupsSettings.groups.patch}},
  \code{\link{groupsSettings.groups.update}}
}
