% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupssettings_functions.R
\name{groupsSettings.groups.get}
\alias{groupsSettings.groups.get}
\title{Gets one resource by id.}
\usage{
groupsSettings.groups.get(groupUniqueId)
}
\arguments{
\item{groupUniqueId}{The resource ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/apps.groups.settings
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/apps.groups.settings)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/google-apps/groups-settings/get_started}{Google Documentation}
}

