% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupssettings_functions.R
\name{groupsSettings.groups.patch}
\alias{groupsSettings.groups.patch}
\title{Updates an existing resource. This method supports patch semantics.}
\usage{
groupsSettings.groups.patch(Groups, groupUniqueId)
}
\arguments{
\item{Groups}{The \link{Groups} object to pass to this method}

\item{groupUniqueId}{The resource ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/apps.groups.settings
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/apps.groups.settings)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/google-apps/groups-settings/get_started}{Google Documentation}

Other Groups.functions: \code{\link{Groups}},
  \code{\link{groupsSettings.groups.update}}
}

