% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupssettings_functions.R
\docType{package}
\name{groupssettings_googleAuthR}
\alias{groupssettings_googleAuthR}
\alias{groupssettings_googleAuthR-package}
\title{Groups Settings API
Lets you manage permission levels and related settings of a group.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:15:26
filename: /Users/mark/dev/R/autoGoogleAPI/googlegroupssettingsv1.auto/R/groupssettings_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/apps.groups.settings
}
}

